requirements.txt
# Core dependencies
google-genai>=1.27.0
python-dotenv>=1.0.0
PyYAML>=6.0
rich>=14.1.0
nest-asyncio>=1.6.0
typing-extensions>=4.0.0
